+++ src/main/java/uk/gov/hmcts/reform/preapi/services/EditRequestService.java
@@ -1,7 +1,5 @@
 package uk.gov.hmcts.reform.preapi.services;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.opencsv.bean.CsvToBeanBuilder;
 import lombok.Cleanup;
 import lombok.extern.slf4j.Slf4j;
@@ -20,7 +18,6 @@ import uk.gov.hmcts.reform.preapi.dto.CreateEditRequestDTO;
 import uk.gov.hmcts.reform.preapi.dto.CreateRecordingDTO;
 import uk.gov.hmcts.reform.preapi.dto.EditCutInstructionDTO;
 import uk.gov.hmcts.reform.preapi.dto.EditRequestDTO;
-import uk.gov.hmcts.reform.preapi.dto.FfmpegEditInstructionDTO;
 import uk.gov.hmcts.reform.preapi.dto.RecordingDTO;
 import uk.gov.hmcts.reform.preapi.entities.EditRequest;
 import uk.gov.hmcts.reform.preapi.entities.Recording;
@@ -32,11 +29,11 @@ import uk.gov.hmcts.reform.preapi.exception.NotFoundException;
 import uk.gov.hmcts.reform.preapi.exception.ResourceInWrongStateException;
 import uk.gov.hmcts.reform.preapi.exception.UnknownServerException;
 import uk.gov.hmcts.reform.preapi.media.edit.EditInstructions;
-import uk.gov.hmcts.reform.preapi.media.edit.FfmpegService;
 import uk.gov.hmcts.reform.preapi.repositories.EditRequestRepository;
 import uk.gov.hmcts.reform.preapi.repositories.RecordingRepository;
 import uk.gov.hmcts.reform.preapi.security.authentication.UserAuthentication;
 import uk.gov.hmcts.reform.preapi.services.edit.AssetGenerationService;
+import uk.gov.hmcts.reform.preapi.services.edit.IEditingService;
 import uk.gov.hmcts.reform.preapi.utils.InputSanitizerUtils;
 
 import java.io.BufferedReader;
@@ -44,23 +41,20 @@ import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.sql.Timestamp;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
-import static uk.gov.hmcts.reform.preapi.dto.EditCutInstructionDTO.formatTime;
 import static uk.gov.hmcts.reform.preapi.media.edit.EditInstructions.fromJson;
+import static uk.gov.hmcts.reform.preapi.utils.JsonUtils.toJson;
 
 @Slf4j
 @Service
-@SuppressWarnings({"PMD.CouplingBetweenObjects", "PMD.GodClass", "PMD.CyclomaticComplexity"})
+@SuppressWarnings({"PMD.CouplingBetweenObjects", "PMD.GodClass"})
 public class EditRequestService {
     private final EditRequestRepository editRequestRepository;
     private final RecordingRepository recordingRepository;
-    private final FfmpegService ffmpegService;
+    private final IEditingService editingService;
     private final RecordingService recordingService;
     private final AssetGenerationService assetGenerationService;
     private final EditNotificationService editNotificationService;
@@ -68,13 +62,13 @@ public class EditRequestService {
     @Autowired
     public EditRequestService(final EditRequestRepository editRequestRepository,
                               final RecordingRepository recordingRepository,
-                              final FfmpegService ffmpegService,
+                              final IEditingService editingService,
                               final RecordingService recordingService,
                               final AssetGenerationService assetGenerationService,
                               final EditNotificationService editNotificationService) {
         this.editRequestRepository = editRequestRepository;
         this.recordingRepository = recordingRepository;
-        this.ffmpegService = ffmpegService;
+        this.editingService = editingService;
         this.recordingService = recordingService;
         this.assetGenerationService = assetGenerationService;
         this.editNotificationService = editNotificationService;
@@ -144,7 +138,7 @@ public class EditRequestService {
         UUID newRecordingId = UUID.randomUUID();
         String filename;
         try {
-            ffmpegService.performEdit(newRecordingId, request);
+            editingService.performEdit(newRecordingId, request);
             filename = assetGenerationService.generateAsset(newRecordingId, request);
         } catch (Exception e) {
             updateEditRequestStatus(request.getId(), EditRequestStatus.ERROR);
@@ -204,7 +198,7 @@ public class EditRequestService {
             return emptyInstructionResult;
         }
 
-        EditRequest request = getEditRequestToCreateOrUpdate(dto, sourceRecording,
+        EditRequest request = editingService.prepareEditRequestToCreateOrUpdate(dto, sourceRecording,
                                                              existingEditRequest.orElse(new EditRequest()));
 
         setCreatedByForNewRequest(request, isUpdate);
@@ -216,6 +210,13 @@ public class EditRequestService {
     @Transactional
     @PreAuthorize("@authorisationService.hasRecordingAccess(authentication, #sourceRecordingId)")
     public EditRequestDTO upsert(UUID sourceRecordingId, MultipartFile file) {
+        if (file.isEmpty()) {
+            throw new BadRequestException("File is empty");
+        }
+        if (file.getContentType() == null || !file.getContentType().equals("text/csv")) {
+            throw new BadRequestException("File type is not supported: expected text/csv");
+        }
+
         // temporary code for create edit request with csv endpoint
         UUID id = UUID.randomUUID();
         CreateEditRequestDTO dto = new CreateEditRequestDTO();
@@ -235,63 +236,6 @@ public class EditRequestService {
         return editRequestRepository.findById(id)
             .map(EditRequestDTO::new)
             .orElseThrow(() -> new UnknownServerException("Edit Request failed to create"));
-    }
-
-    private @NotNull EditRequest getEditRequestToCreateOrUpdate(CreateEditRequestDTO dto, Recording sourceRecording,
-                                                                EditRequest request) {
-        if (dto.isForceReencode()) {
-            request.setId(dto.getId());
-            request.setSourceRecording(sourceRecording);
-            request.setStatus(dto.getStatus());
-            request.setJointlyAgreed(dto.getJointlyAgreed());
-            request.setApprovedAt(dto.getApprovedAt());
-            request.setApprovedBy(dto.getApprovedBy());
-            request.setRejectionReason(dto.getRejectionReason());
-            request.setEditInstruction(toJson(new EditInstructions(List.of(), List.of(), true,
-                                                                   shouldSendNotifications(dto))));
-            return request;
-        }
-
-        boolean isOriginalRecordingEdit = sourceRecording.getParentRecording() == null;
-
-        boolean sourceInstructionsAreNotEmpty = !isOriginalRecordingEdit
-            && sourceRecording.getEditInstruction() != null
-            && !sourceRecording.getEditInstruction().isEmpty();
-
-        EditInstructions prevInstructions = null;
-        if (sourceInstructionsAreNotEmpty) {
-            prevInstructions = EditInstructions.tryFromJson(sourceRecording.getEditInstruction());
-        }
-        boolean prevInstructionsAreNotEmpty = prevInstructions != null
-            && prevInstructions.getFfmpegInstructions() != null
-            && !prevInstructions.getFfmpegInstructions().isEmpty()
-            && prevInstructions.getRequestedInstructions() != null
-            && !prevInstructions.getRequestedInstructions().isEmpty();
-
-        boolean isInstructionCombination = sourceInstructionsAreNotEmpty && prevInstructionsAreNotEmpty;
-
-        request.setId(dto.getId());
-        request.setSourceRecording(!isInstructionCombination
-                                       ? sourceRecording
-                                       : sourceRecording.getParentRecording());
-        request.setStatus(dto.getStatus());
-        request.setJointlyAgreed(dto.getJointlyAgreed());
-        request.setApprovedAt(dto.getApprovedAt());
-        request.setApprovedBy(dto.getApprovedBy());
-        request.setRejectionReason(dto.getRejectionReason());
-
-        List<EditCutInstructionDTO> requestedEdits = isInstructionCombination
-            ? combineCutsOnOriginalTimeline(prevInstructions, dto.getEditInstructions())
-            : dto.getEditInstructions();
-
-        List<FfmpegEditInstructionDTO> editInstructions = invertInstructions(
-            requestedEdits,
-            isInstructionCombination ? request.getSourceRecording() : sourceRecording
-        );
-
-        request.setEditInstruction(toJson(new EditInstructions(requestedEdits, editInstructions, false,
-                                                               shouldSendNotifications(dto))));
-        return request;
     }
 
     private void validateEditMode(CreateEditRequestDTO dto) {
@@ -359,10 +303,6 @@ public class EditRequestService {
         editNotificationService.onEditRequestRejected(request);
     }
 
-    private boolean shouldSendNotifications(CreateEditRequestDTO dto) {
-        return !Boolean.FALSE.equals(dto.getSendNotifications());
-    }
-
     private List<EditCutInstructionDTO> parseCsv(MultipartFile file) {
         try {
             @Cleanup BufferedReader reader = new BufferedReader(new InputStreamReader(
@@ -376,158 +316,6 @@ public class EditRequestService {
         } catch (Exception e) {
             log.error("Error when reading CSV file: {} ", e.getMessage());
             throw new UnknownServerException("Uploaded CSV file incorrectly formatted", e);
-        }
-    }
-
-    @SuppressWarnings({"PMD.CognitiveComplexity", "PMD.AvoidLiteralsInIfCondition"})
-    public List<FfmpegEditInstructionDTO> invertInstructions(final List<EditCutInstructionDTO> instructions,
-                                                             final Recording recording) {
-        long recordingDuration = recording.getDuration().toSeconds();
-        if (instructions.size() == 1) {
-            EditCutInstructionDTO firstInstruction = instructions.getFirst();
-            if (firstInstruction.getStart() == 0 && firstInstruction.getEnd() == recordingDuration) {
-                throw new BadRequestException("Invalid Instruction: Cannot cut an entire recording: Start("
-                                                  + formatTime(firstInstruction.getStart())
-                                                  + "), End("
-                                                  + formatTime(firstInstruction.getEnd())
-                                                  + "), Recording Duration("
-                                                  + formatTime(recordingDuration)
-                                                  + ")");
-            }
-        }
-
-        instructions.sort(Comparator.comparing(EditCutInstructionDTO::getStart)
-                              .thenComparing(EditCutInstructionDTO::getEnd));
-
-        for (int i = 1; i < instructions.size(); i++) {
-            EditCutInstructionDTO prev = instructions.get(i - 1);
-            EditCutInstructionDTO curr = instructions.get(i);
-            if (curr.getStart() < prev.getEnd()) {
-                throw new BadRequestException("Overlapping instructions: Previous End("
-                                                  + formatTime(prev.getEnd())
-                                                  + "), Current Start("
-                                                  + formatTime(curr.getStart())
-                                                  + ")");
-            }
-        }
-
-        long currentTime = 0L;
-        List<FfmpegEditInstructionDTO> invertedInstructions = new ArrayList<>();
-
-        // invert
-        for (EditCutInstructionDTO instruction : instructions) {
-            if (instruction.getStart() == instruction.getEnd()) {
-                throw new BadRequestException(
-                    "Invalid instruction: Instruction with 0 second duration invalid: Start("
-                        + formatTime(instruction.getStart())
-                        + "), End("
-                        + formatTime(instruction.getEnd())
-                        + ")");
-            }
-            if (instruction.getEnd() < instruction.getStart()) {
-                throw new BadRequestException(
-                    "Invalid instruction: Instruction with end time before start time: Start("
-                        + formatTime(instruction.getStart())
-                        + "), End("
-                        + formatTime(instruction.getEnd())
-                        + ")");
-            }
-            if (instruction.getEnd() > recordingDuration) {
-                throw new BadRequestException("Invalid instruction: Instruction end time exceeding duration: Start("
-                                                  + formatTime(instruction.getStart())
-                                                  + "), End("
-                                                  + formatTime(instruction.getEnd())
-                                                  + "), Recording Duration("
-                                                  + formatTime(recordingDuration)
-                                                  + ")");
-            }
-            if (currentTime < instruction.getStart()) {
-                invertedInstructions.add(new FfmpegEditInstructionDTO(currentTime, instruction.getStart()));
-            }
-            currentTime = Math.max(currentTime, instruction.getEnd());
-        }
-        if (currentTime != recordingDuration) {
-            invertedInstructions.add(new FfmpegEditInstructionDTO(currentTime, recordingDuration));
-        }
-
-        return invertedInstructions;
-    }
-
-    protected List<EditCutInstructionDTO> combineCutsOnOriginalTimeline(
-        final EditInstructions original,
-        final List<EditCutInstructionDTO> newInstructions
-    ) {
-        final List<FfmpegEditInstructionDTO> keptSegments = original.getFfmpegInstructions();
-
-        final List<FfmpegEditInstructionDTO> editedTimelineMapping = new ArrayList<>();
-        long cursor = 0;
-        for (FfmpegEditInstructionDTO segment : keptSegments) {
-            long duration = segment.getEnd() - segment.getStart();
-            editedTimelineMapping.add(new FfmpegEditInstructionDTO(cursor, cursor + duration));
-            cursor += duration;
-        }
-
-        final List<EditCutInstructionDTO> mappedCuts = new ArrayList<>();
-        for (EditCutInstructionDTO newCut : newInstructions) {
-            for (int i = 0; i < keptSegments.size(); i++) {
-                final FfmpegEditInstructionDTO originalSegment = keptSegments.get(i);
-                final FfmpegEditInstructionDTO editedSegment = editedTimelineMapping.get(i);
-
-                long start = editedSegment.getStart();
-                long end = editedSegment.getEnd();
-
-                if (newCut.getEnd() <= start || newCut.getStart() >= end) {
-                    continue;
-                }
-
-                long overlapStart = Math.max(start, newCut.getStart());
-                long overlapEnd = Math.min(end, newCut.getEnd());
-
-                long offsetInSegment = overlapStart - start;
-                long cutLength = overlapEnd - overlapStart;
-
-                long originalMappedStart = originalSegment.getStart() + offsetInSegment;
-                long originalMappedEnd = originalMappedStart + cutLength;
-
-                mappedCuts.add(new EditCutInstructionDTO(
-                    originalMappedStart,
-                    originalMappedEnd,
-                    newCut.getReason()
-                ));
-            }
-        }
-
-        mappedCuts.addAll(original.getRequestedInstructions());
-        mappedCuts.sort(Comparator.comparing(EditCutInstructionDTO::getStart));
-
-        return mergeOverlappingCuts(mappedCuts)
-            .stream()
-            .map(cut -> new EditCutInstructionDTO(cut.getStart(), cut.getEnd(), cut.getReason()))
-            .collect(Collectors.toList());
-    }
-
-    protected List<EditCutInstructionDTO> mergeOverlappingCuts(final List<EditCutInstructionDTO> cuts) {
-        final List<EditCutInstructionDTO> merged = new ArrayList<>();
-        EditCutInstructionDTO current = cuts.getFirst();
-
-        for (int i = 1; i < cuts.size(); i++) {
-            final EditCutInstructionDTO next = cuts.get(i);
-            if (next.getStart() <= current.getEnd()) {
-                current.setEnd(Math.max(current.getEnd(), next.getEnd()));
-            } else {
-                merged.add(current);
-                current = next;
-            }
-        }
-        merged.add(current);
-        return merged;
-    }
-
-    protected <E> String toJson(E instructions) {
-        try {
-            return new ObjectMapper().writeValueAsString(instructions);
-        } catch (JsonProcessingException e) {
-            throw new UnknownServerException("Something went wrong: " + e.getMessage(), e);
         }
     }
