+++ src/test/java/uk/gov/hmcts/reform/preapi/services/edit/FfmpegServiceTest.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.preapi.media.edit;
+package uk.gov.hmcts.reform.preapi.services.edit;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.exec.CommandLine;
+import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,12 +11,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import uk.gov.hmcts.reform.preapi.component.CommandExecutor;
+import uk.gov.hmcts.reform.preapi.dto.CreateEditRequestDTO;
+import uk.gov.hmcts.reform.preapi.dto.EditCutInstructionDTO;
 import uk.gov.hmcts.reform.preapi.dto.FfmpegEditInstructionDTO;
 import uk.gov.hmcts.reform.preapi.entities.EditRequest;
 import uk.gov.hmcts.reform.preapi.entities.Recording;
 import uk.gov.hmcts.reform.preapi.enums.EditRequestStatus;
+import uk.gov.hmcts.reform.preapi.exception.BadRequestException;
 import uk.gov.hmcts.reform.preapi.exception.NotFoundException;
 import uk.gov.hmcts.reform.preapi.exception.UnknownServerException;
+import uk.gov.hmcts.reform.preapi.media.edit.EditInstructions;
 import uk.gov.hmcts.reform.preapi.media.storage.AzureFinalStorageService;
 import uk.gov.hmcts.reform.preapi.media.storage.AzureIngestStorageService;
 
@@ -23,6 +28,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.time.Duration;
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -37,6 +43,8 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.reform.preapi.services.edit.EditRequestTestUtils.assertEditInstructionsEq;
+import static uk.gov.hmcts.reform.preapi.utils.JsonUtils.toJson;
 
 @SpringBootTest(classes = FfmpegService.class)
 public class FfmpegServiceTest {
@@ -50,21 +58,29 @@ public class FfmpegServiceTest {
     private CommandExecutor commandExecutor;
 
     @Autowired
-    private FfmpegService ffmpegService;
+    private FfmpegService underTest;
 
-    private Recording recording;
+    @MockitoBean
+    private Recording mockRecording;
+
+    @MockitoBean
+    private Recording mockParentRecording;
+
     private EditRequest editRequest;
 
     @BeforeEach
     void setUp() {
-        recording = new Recording();
-        recording.setId(UUID.randomUUID());
-        recording.setFilename("input.mp4");
-        recording.setDuration(Duration.ofMinutes(3));
+        when(mockRecording.getId()).thenReturn(UUID.randomUUID());
+        when(mockRecording.getFilename()).thenReturn("input.mp4");
+        when(mockRecording.getDuration()).thenReturn(Duration.ofMinutes(3));
+
+        when(mockParentRecording.getId()).thenReturn(UUID.randomUUID());
+        when(mockParentRecording.getFilename()).thenReturn("parent.mp4");
+        when(mockParentRecording.getDuration()).thenReturn(Duration.ofMinutes(30));
 
         editRequest = new EditRequest();
         editRequest.setId(UUID.randomUUID());
-        editRequest.setSourceRecording(recording);
+        editRequest.setSourceRecording(mockRecording);
         editRequest.setEditInstruction("{}");
         editRequest.setStatus(EditRequestStatus.PENDING);
     }
@@ -72,12 +88,12 @@ public class FfmpegServiceTest {
     @Test
     @DisplayName("Should throw error when source recording does not have a filename set")
     void performEditFileNameNull() {
-        recording.setFilename(null);
-        var newRecordingId = UUID.randomUUID();
+        when(mockRecording.getFilename()).thenReturn(null);
+        UUID newRecordingId = UUID.randomUUID();
 
-        var message = assertThrows(
+        String message = assertThrows(
             NotFoundException.class,
-            () -> ffmpegService.performEdit(newRecordingId, editRequest)
+            () -> underTest.performEdit(newRecordingId, editRequest)
         ).getMessage();
 
         assertThat(message).isEqualTo("Not found: No file name provided");
@@ -91,16 +107,16 @@ public class FfmpegServiceTest {
     @DisplayName("Should throw error when source recording does not have a filename set")
     void performEditDownloadFailure() throws JsonProcessingException {
         setRandomEditInstructions();
-        var newRecordingId = UUID.randomUUID();
-        var inputFileName = editRequest.getSourceRecording().getFilename();
-        var containerName = editRequest.getSourceRecording().getId().toString();
+        UUID newRecordingId = UUID.randomUUID();
+        String inputFileName = editRequest.getSourceRecording().getFilename();
+        String containerName = editRequest.getSourceRecording().getId().toString();
 
         when(azureFinalStorageService.downloadBlob(containerName, inputFileName, inputFileName))
             .thenReturn(false);
 
-        var message = assertThrows(
+        String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.performEdit(newRecordingId, editRequest)
+            () -> underTest.performEdit(newRecordingId, editRequest)
         ).getMessage();
 
         assertThat(message)
@@ -116,17 +132,17 @@ public class FfmpegServiceTest {
     @DisplayName("Should throw error when ffmpeg execution encounters and error")
     void performEditExecuteFailure() throws JsonProcessingException {
         setRandomEditInstructions();
-        var newRecordingId = UUID.randomUUID();
-        var inputFileName = editRequest.getSourceRecording().getFilename();
-        var containerName = editRequest.getSourceRecording().getId().toString();
+        UUID newRecordingId = UUID.randomUUID();
+        String inputFileName = editRequest.getSourceRecording().getFilename();
+        String containerName = editRequest.getSourceRecording().getId().toString();
 
         when(azureFinalStorageService.downloadBlob(containerName, inputFileName, inputFileName))
             .thenReturn(true);
         when(commandExecutor.execute(any(CommandLine.class))).thenReturn(false);
 
-        var message = assertThrows(
+        String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.performEdit(newRecordingId, editRequest)
+            () -> underTest.performEdit(newRecordingId, editRequest)
         ).getMessage();
 
         assertThat(message)
@@ -149,7 +165,7 @@ public class FfmpegServiceTest {
             .thenReturn(true);
         when(commandExecutor.executeAndGetOutput(any(CommandLine.class))).thenReturn(mockOutput);
 
-        Duration duration = ffmpegService.getDurationFromMp4(containerName, fileName);
+        Duration duration = underTest.getDurationFromMp4(containerName, fileName);
 
         assertThat(duration).isNotNull();
         assertThat(duration.toMillis()).isEqualTo(123456);
@@ -164,7 +180,7 @@ public class FfmpegServiceTest {
             .thenReturn(true);
         when(commandExecutor.executeAndGetOutput(any(CommandLine.class))).thenReturn("not_a_number");
 
-        Duration duration = ffmpegService.getDurationFromMp4(containerName, fileName);
+        Duration duration = underTest.getDurationFromMp4(containerName, fileName);
 
         assertThat(duration).isNull();
     }
@@ -179,7 +195,7 @@ public class FfmpegServiceTest {
         when(commandExecutor.executeAndGetOutput(any(CommandLine.class)))
             .thenThrow(new RuntimeException("An error occurred"));
 
-        Duration duration = ffmpegService.getDurationFromMp4(containerName, fileName);
+        Duration duration = underTest.getDurationFromMp4(containerName, fileName);
 
         assertThat(duration).isNull();
     }
@@ -193,7 +209,7 @@ public class FfmpegServiceTest {
             .thenReturn(true);
         when(commandExecutor.executeAndGetOutput(any(CommandLine.class))).thenReturn("  ");
 
-        Duration duration = ffmpegService.getDurationFromMp4(containerName, fileName);
+        Duration duration = underTest.getDurationFromMp4(containerName, fileName);
 
         assertThat(duration).isNull();
     }
@@ -205,7 +221,7 @@ public class FfmpegServiceTest {
 
         String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.generateMultiEditCommands(editRequest, "", "")
+            () -> underTest.generateMultiEditCommands(editRequest, "", "")
         ).getMessage();
 
         assertThat(message).isEqualTo("Unknown Server Exception: Unable to read edit instructions");
@@ -216,7 +232,7 @@ public class FfmpegServiceTest {
     void generateCommandFromJsonEmpty() {
         String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.generateMultiEditCommands(editRequest, "", "")
+            () -> underTest.generateMultiEditCommands(editRequest, "", "")
         ).getMessage();
 
         assertThat(message).isEqualTo("Unknown Server Exception: Malformed edit instructions");
@@ -228,7 +244,7 @@ public class FfmpegServiceTest {
         editRequest.setEditInstruction("{\"ffmpegInstructions\": []}");
         String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.generateMultiEditCommands(editRequest, "", "")
+            () -> underTest.generateMultiEditCommands(editRequest, "", "")
         ).getMessage();
 
         assertThat(message).isEqualTo("Unknown Server Exception: Malformed edit instructions");
@@ -240,7 +256,7 @@ public class FfmpegServiceTest {
         List<FfmpegEditInstructionDTO> instructions = setRandomEditInstructions();
 
         LinkedHashMap<String, CommandLine>
-            commands = ffmpegService.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
+            commands = underTest.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
 
         assertThat(commands).isNotNull();
         assertThat(commands).hasSize(2);
@@ -290,7 +306,7 @@ public class FfmpegServiceTest {
     void generateSingleEditCommandInMultiCommandMode() {
         editRequest.setEditInstruction("{\"ffmpegInstructions\": [{\"start\": 10, \"end\": 20}]}");
         LinkedHashMap<String, CommandLine> commands =
-            ffmpegService.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
+            underTest.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
 
         assertThat(commands).hasSize(1);
         assertThat(commands).containsKey("output.mp4");
@@ -306,7 +322,7 @@ public class FfmpegServiceTest {
         );
 
         LinkedHashMap<String, CommandLine> commands =
-            ffmpegService.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
+            underTest.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
 
         assertThat(commands).hasSize(1);
         assertThat(commands).containsKey("output.mp4");
@@ -330,7 +346,7 @@ public class FfmpegServiceTest {
     void generateMultipleEditCommandsAndConcatenateSegments() throws JsonProcessingException {
         setRandomEditInstructions();
         LinkedHashMap<String, CommandLine> commands =
-            ffmpegService.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
+            underTest.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
 
         assertThat(commands).hasSizeGreaterThan(1);
 
@@ -351,7 +367,7 @@ public class FfmpegServiceTest {
 
         String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.performEdit(newRecordingId, editRequest)
+            () -> underTest.performEdit(newRecordingId, editRequest)
         ).getMessage();
 
         assertThat(message).isEqualTo(
@@ -364,7 +380,7 @@ public class FfmpegServiceTest {
     @DisplayName("Should successfully generate the concat list file")
     void successfullyGenerateConcatListFile() throws IOException {
         Set<String> segmentFiles = Set.of("segment1.mp4", "segment2.mp4");
-        ffmpegService.generateConcatListFile(segmentFiles, "concat-list.txt");
+        underTest.generateConcatListFile(segmentFiles, "concat-list.txt");
 
         File outputFile = new File("concat-list.txt");
         assertThat(outputFile).exists();
@@ -382,7 +398,7 @@ public class FfmpegServiceTest {
         when(azureIngestStorageService.uploadBlob(any(), any(), any())).thenReturn(true);
 
         UUID newRecordingId = UUID.randomUUID();
-        ffmpegService.performEdit(newRecordingId, editRequest);
+        underTest.performEdit(newRecordingId, editRequest);
 
         // 2 segment commands + 1 concat command = 3 ffmpeg executions
         verify(commandExecutor, times(3)).execute(any(CommandLine.class));
@@ -405,7 +421,7 @@ public class FfmpegServiceTest {
 
         String message = assertThrows(
             UnknownServerException.class,
-            () -> ffmpegService.performEdit(newRecordingId, editRequest)
+            () -> underTest.performEdit(newRecordingId, editRequest)
         ).getMessage();
 
         assertThat(message)
@@ -419,13 +435,13 @@ public class FfmpegServiceTest {
     @Test
     @DisplayName("Should pass correct segment file names to concat list generator")
     void shouldPassCorrectFilesToConcatListFile() throws IOException {
-        var instructions = setRandomEditInstructions();
-        var commands = ffmpegService.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
+        List<FfmpegEditInstructionDTO> instructions = setRandomEditInstructions();
+        LinkedHashMap<String, CommandLine> commands = underTest.generateMultiEditCommands(editRequest, "input.mp4", "output.mp4");
 
-        var concatListFile = new File("test-concat-list.txt");
-        ffmpegService.generateConcatListFile(commands.keySet(), concatListFile.getName());
+        File concatListFile = new File("test-concat-list.txt");
+        underTest.generateConcatListFile(commands.keySet(), concatListFile.getName());
 
-        var lines = Files.readAllLines(concatListFile.toPath());
+        List<@NotNull String> lines = Files.readAllLines(concatListFile.toPath());
         assertThat(lines).hasSize(instructions.size());
         lines.forEach(line -> assertThat(line).startsWith("file '").endsWith(".mp4'"));
 
@@ -442,7 +458,7 @@ public class FfmpegServiceTest {
         when(azureIngestStorageService.uploadBlob(any(), any(), any())).thenReturn(true);
 
         UUID newRecordingId = UUID.randomUUID();
-        ffmpegService.performEdit(newRecordingId, editRequest);
+        underTest.performEdit(newRecordingId, editRequest);
 
         // Only one command executed (no concat)
         verify(commandExecutor, times(1)).execute(any());
@@ -463,24 +479,301 @@ public class FfmpegServiceTest {
         when(azureIngestStorageService.uploadBlob(any(), any(), any())).thenReturn(true);
 
         UUID newRecordingId = UUID.randomUUID();
-        ffmpegService.performEdit(newRecordingId, editRequest);
+        underTest.performEdit(newRecordingId, editRequest);
 
         verify(commandExecutor, times(1)).execute(any());
         verify(azureFinalStorageService, times(1)).downloadBlob(any(), any(), any());
         verify(azureIngestStorageService, times(1)).uploadBlob(any(), any(), any());
     }
 
+    @Test
+    @DisplayName("Should throw bad request when instruction cuts entire recording")
+    void invertInstructionsBadRequestCutToZeroDuration() {
+        List<EditCutInstructionDTO> instructions = new ArrayList<>();
+        instructions.add(EditCutInstructionDTO.builder()
+            .start(0L)
+            .end(180L)
+            .build());
+
+        String message = assertThrows(
+            BadRequestException.class,
+            () -> underTest.invertInstructions(instructions, mockRecording)
+        ).getMessage();
+
+        assertThat(message)
+            .isEqualTo("Invalid Instruction: Cannot cut an entire recording: "
+                + "Start(00:00:00), End(00:03:00), "
+                + "Recording Duration(00:03:00)");
+    }
+
+    @Test
+    @DisplayName("Should throw bad request when instruction has same value for start and end")
+    void invertInstructionsBadRequestStartEndEqual() {
+        List<EditCutInstructionDTO> instructions = new ArrayList<>();
+        instructions.add(EditCutInstructionDTO.builder()
+            .start(60L)
+            .end(60L)
+            .build());
+
+        String message = assertThrows(
+            BadRequestException.class,
+            () -> underTest.invertInstructions(instructions, mockRecording)
+        ).getMessage();
+
+        assertThat(message)
+            .isEqualTo("Invalid instruction: Instruction with 0 second duration invalid: "
+                + "Start(00:01:00), End(00:01:00)");
+    }
+
+    @Test
+    @DisplayName("Should throw bad request when instruction end time is less than start time")
+    void invertInstructionsBadRequestEndLTStart() {
+        List<EditCutInstructionDTO> instructions = new ArrayList<>();
+        instructions.add(EditCutInstructionDTO.builder()
+            .start(60L)
+            .end(50L)
+            .build());
+
+        String message = assertThrows(
+            BadRequestException.class,
+            () -> underTest.invertInstructions(instructions, mockRecording)
+        ).getMessage();
+
+        assertThat(message)
+            .isEqualTo("Invalid instruction: Instruction with end time before start time: "
+                + "Start(00:01:00), End(00:00:50)");
+    }
+
+    @Test
+    @DisplayName("Should throw bad request when instruction end time exceeds duration")
+    void invertInstructionsBadRequestEndTimeExceedsDuration() {
+        List<EditCutInstructionDTO> instructions = new ArrayList<>();
+        instructions.add(EditCutInstructionDTO.builder()
+            .start(60L)
+            .end(200L) // duration is 180
+            .build());
+
+        String message = assertThrows(
+            BadRequestException.class,
+            () -> underTest.invertInstructions(instructions, mockRecording)
+        ).getMessage();
+
+        assertThat(message)
+            .isEqualTo("Invalid instruction: Instruction end time exceeding duration: "
+                + "Start(00:01:00), End(00:03:20), "
+                + "Recording Duration(00:03:00)");
+    }
+
+    @Test
+    @DisplayName("Should throw bad request when instructions overlap")
+    void invertInstructionsOverlap() {
+        List<EditCutInstructionDTO> instructions = new ArrayList<>();
+        instructions.add(EditCutInstructionDTO.builder()
+            .start(10L)
+            .end(30L)
+            .build());
+        instructions.add(EditCutInstructionDTO.builder()
+            .start(20L)
+            .end(40L)
+            .build());
+
+        String message = assertThrows(
+            BadRequestException.class,
+            () -> underTest.invertInstructions(instructions, mockRecording)
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Overlapping instructions: "
+            + "Previous End(00:00:30), Current Start(00:00:20)");
+    }
+
+    @Test
+    @DisplayName("Should return inverted instructions (ordered correctly)")
+    void invertInstructionsSuccess() {
+        List<EditCutInstructionDTO> instructions1 = new ArrayList<>();
+        instructions1.add(EditCutInstructionDTO.builder()
+            .start(60L)
+            .end(120L)
+            .build());
+        instructions1.add(EditCutInstructionDTO.builder()
+            .start(150L)
+            .end(180L)
+            .build());
+
+        List<@NotNull FfmpegEditInstructionDTO> expectedInvertedInstructions = List.of(
+            FfmpegEditInstructionDTO.builder()
+                .start(0)
+                .end(60)
+                .build(),
+            FfmpegEditInstructionDTO.builder()
+                .start(120)
+                .end(150)
+                .build()
+        );
+
+        assertEditInstructionsEq(expectedInvertedInstructions,
+            underTest.invertInstructions(instructions1, mockRecording));
+    }
+
+    @Test
+    @DisplayName("Should return inverted instructions (ordered correctly) when not cutting the end")
+    void invertInstructionsNotCuttingEndSuccess() {
+        List<EditCutInstructionDTO> instructions1 = new ArrayList<>();
+        instructions1.add(EditCutInstructionDTO.builder()
+            .start(60L)
+            .end(120L)
+            .build());
+        instructions1.add(EditCutInstructionDTO.builder()
+            .start(150L)
+            .end(160L)
+            .build());
+
+        List<@NotNull FfmpegEditInstructionDTO> expectedInvertedInstructions = List.of(
+            FfmpegEditInstructionDTO.builder()
+                .start(0)
+                .end(60)
+                .build(),
+            FfmpegEditInstructionDTO.builder()
+                .start(120)
+                .end(150)
+                .build(),
+            FfmpegEditInstructionDTO.builder()
+                .start(160)
+                .end(180)
+                .build()
+        );
+
+        assertEditInstructionsEq(expectedInvertedInstructions,
+            underTest.invertInstructions(instructions1, mockRecording));
+    }
+
+    @Test
+    @DisplayName("Should combine instructions correctly with single new command")
+    void combineCutsOnOriginalTimelineSingleCommand() {
+        List<@NotNull FfmpegEditInstructionDTO> originallyKeptSegments = List.of(createSegment(0, 10), createSegment(20, 30));
+        List<@NotNull EditCutInstructionDTO> originalCutSegments = List.of(createCut(10, 20, "some reason"));
+        EditInstructions originalInstructions = new EditInstructions(originalCutSegments, originallyKeptSegments);
+
+        // Cut at 5–8 in the edited timeline mapping to 5–8 in the original timeline
+        EditCutInstructionDTO newCut = createCut(5, 8, "test");
+
+        List<EditCutInstructionDTO> result = underTest.combineCutsOnOriginalTimeline(originalInstructions, List.of(newCut));
+
+        assertThat(result).hasSize(2);
+
+        assertThat(result.getFirst().getStart()).isEqualTo(5);
+        assertThat(result.getFirst().getEnd()).isEqualTo(8);
+        assertThat(result.getFirst().getReason()).isEqualTo("test");
+
+        assertThat(result.getLast().getStart()).isEqualTo(10);
+        assertThat(result.getLast().getEnd()).isEqualTo(20);
+        assertThat(result.getLast().getReason()).isEqualTo("some reason");
+    }
+
+    @Test
+    @DisplayName("Should combine instructions correctly with cuts across multiple segments")
+    void combineCutsOnOriginalTimelineMapsCutThatSpanningMultipleSegments() {
+        List<@NotNull FfmpegEditInstructionDTO> originallyKeptSegments = List.of(createSegment(0, 10), createSegment(20, 30));
+        List<@NotNull EditCutInstructionDTO> originalCutSegments = List.of(createCut(10, 20, "some reason"));
+        EditInstructions originalInstructions = new EditInstructions(originalCutSegments, originallyKeptSegments);
+
+        // Cut at 8–12 in the edited timeline:
+        // - 8–10 -> 8–10 in original (segment 1)
+        // - 10–12 -> 20–22 in original (segment 2)
+        EditCutInstructionDTO newCut = createCut(8, 12, "test");
+
+        List<EditCutInstructionDTO> result = underTest.combineCutsOnOriginalTimeline(originalInstructions, List.of(newCut));
+
+        assertThat(result).hasSize(1);
+
+        assertThat(result.getFirst().getStart()).isEqualTo(8);
+        assertThat(result.getFirst().getEnd()).isEqualTo(22);
+        assertThat(result.getFirst().getReason()).isEqualTo("test");
+    }
+
+    @Test
+    @DisplayName("Should upsert when isOriginalRecordingEdit is false and isInstructionCombination false")
+    void upsertIsOriginalRecordingEditFalseIsInstructionCombinationFalse() {
+        // when editing an edit from legacy editing
+        when(mockRecording.getFilename()).thenReturn("filename.mp4");
+        when(mockRecording.getDuration()).thenReturn(Duration.ofSeconds(30));
+
+        CreateEditRequestDTO request = new CreateEditRequestDTO();
+        request.setId(UUID.randomUUID());
+        request.setSourceRecordingId(mockRecording.getId());
+        request.setStatus(EditRequestStatus.PENDING);
+        request.setEditInstructions(new ArrayList<>(List.of(createCut(10, 20, "some reason"))));
+
+        EditRequest result = underTest.prepareEditRequestToCreateOrUpdate(request, mockRecording, editRequest);
+        assertThat(result.getId()).isEqualTo(request.getId());
+        assertThat(result.getSourceRecording().getId()).isEqualTo(mockRecording.getId());
+        assertThat(result.getStatus()).isEqualTo(EditRequestStatus.PENDING);
+        assertThat(result.getEditInstruction()).isNotNull();
+
+        EditInstructions editInstructions = EditInstructions.tryFromJson(result.getEditInstruction());
+        assertThat(editInstructions).isNotNull();
+        assertThat(editInstructions.getRequestedInstructions()).isNotNull();
+        assertThat(editInstructions.getRequestedInstructions()).hasSize(1);
+        assertThat(editInstructions.getRequestedInstructions().getFirst().getStart()).isEqualTo(10);
+        assertThat(editInstructions.getRequestedInstructions().getFirst().getEnd()).isEqualTo(20);
+
+        assertThat(editInstructions.getFfmpegInstructions()).isNotNull();
+        assertThat(editInstructions.shouldSendNotifications()).isTrue();
+
+        assertEditInstructionsEq(List.of(createSegment(0, 10), createSegment(20, 30)),
+                editInstructions.getFfmpegInstructions());
+    }
+
+    @Test
+    @DisplayName("Should upsert when isOriginalRecordingEdit is false and isInstructionCombination true")
+    void upsertIsOriginalRecordingEditFalseIsInstructionCombinationTrue() {
+        // when editing an edit from the new editing process
+        final EditInstructions originalEdits = new EditInstructions(
+                List.of(createCut(10, 20, "some original reason")),
+                List.of(createSegment(0, 10), createSegment(20, 30)));
+
+        CreateEditRequestDTO request = new CreateEditRequestDTO();
+        request.setId(UUID.randomUUID());
+        request.setSourceRecordingId(mockRecording.getId());
+        request.setStatus(EditRequestStatus.PENDING);
+        request.setEditInstructions(new ArrayList<>(List.of(createCut(5, 8, "some new reason"))));
+
+        when(mockRecording.getParentRecording()).thenReturn(mockParentRecording);
+        when(mockRecording.getEditInstruction()).thenReturn(toJson(originalEdits));
+
+        EditRequest result = underTest.prepareEditRequestToCreateOrUpdate(request, mockRecording, editRequest);
+        assertThat(result.getId()).isEqualTo(request.getId());
+        assertThat(result.getSourceRecording().getId()).isEqualTo(mockParentRecording.getId());
+        assertThat(result.getStatus()).isEqualTo(EditRequestStatus.PENDING);
+        assertThat(result.getEditInstruction()).isNotNull();
+
+        EditInstructions editInstructions = EditInstructions.tryFromJson(result.getEditInstruction());
+        assertThat(editInstructions).isNotNull();
+        assertThat(editInstructions.getRequestedInstructions()).isNotNull();
+        assertThat(editInstructions.getRequestedInstructions()).hasSize(2);
+        assertThat(editInstructions.getRequestedInstructions().getFirst().getStart()).isEqualTo(5);
+        assertThat(editInstructions.getRequestedInstructions().getFirst().getEnd()).isEqualTo(8);
+        assertThat(editInstructions.getRequestedInstructions().getFirst().getReason()).isEqualTo("some new reason");
+        assertThat(editInstructions.getRequestedInstructions().getLast().getStart()).isEqualTo(10);
+        assertThat(editInstructions.getRequestedInstructions().getLast().getEnd()).isEqualTo(20);
+        assertThat(editInstructions.getRequestedInstructions().getLast().getReason()).isEqualTo("some original reason");
+
+        assertThat(editInstructions.getFfmpegInstructions()).isNotNull();
+
+        assertEditInstructionsEq(List.of(createSegment(0, 5), createSegment(8, 10), createSegment(20, 1800)),
+                editInstructions.getFfmpegInstructions());
+    }
+
     private String generateEditInstructionsJson(List<FfmpegEditInstructionDTO> ffmpegEditInstructions)
         throws JsonProcessingException {
-        var objectMapper = new ObjectMapper();
-        var instructions = new EditInstructions(List.of(), ffmpegEditInstructions);
+        ObjectMapper objectMapper = new ObjectMapper();
+        EditInstructions instructions = new EditInstructions(List.of(), ffmpegEditInstructions);
 
         return objectMapper.writeValueAsString(instructions);
     }
 
     private List<FfmpegEditInstructionDTO> setRandomEditInstructions() throws JsonProcessingException {
-        var random = new Random();
-        var instructions = List.of(
+        Random random = new Random();
+        List<@NotNull FfmpegEditInstructionDTO> instructions = List.of(
             FfmpegEditInstructionDTO.builder()
                 .start(random.nextInt(0, 10))
                 .end(random.nextInt(60, 70))
@@ -492,5 +785,13 @@ public class FfmpegServiceTest {
         );
         editRequest.setEditInstruction(generateEditInstructionsJson(instructions));
         return instructions;
+    }
+
+    private static FfmpegEditInstructionDTO createSegment(long start, long end) {
+        return new FfmpegEditInstructionDTO(start, end);
+    }
+
+    private static EditCutInstructionDTO createCut(long start, long end, String reason) {
+        return new EditCutInstructionDTO(start, end, reason);
     }
 }
